Refactor structure of Sync dirs
