+++ src/hooks/sync-sites/use-sync-pull.ts
@@ -3,7 +3,6 @@ import { sprintf } from '@wordpress/i18n';
 import { useI18n } from '@wordpress/react-i18n';
 import { useCallback, useEffect, useMemo } from 'react';
 import { SYNC_PUSH_SIZE_LIMIT_GB, SYNC_PUSH_SIZE_LIMIT_BYTES } from 'src/constants';
-import { SyncOption } from 'src/hooks/sync-sites/sync-option';
 import {
 	ClearState,
 	generateStateId,
@@ -22,6 +21,7 @@ import {
 import { getIpcApi } from 'src/lib/get-ipc-api';
 import { getHostnameFromUrl } from 'src/lib/url-utils';
 import type { SyncSite } from 'src/hooks/use-fetch-wpcom-sites/types';
+import type { SyncOption } from 'src/types';
 
 export type SyncBackupState = {
 	remoteSiteId: number;
