+++ src/modules/content-tab-sync/components/sync-connected-sites.tsx
@@ -15,7 +15,6 @@ import { SyncPullPushClear } from 'src/components/sync-pull-push-clear';
 import { Tooltip, DynamicTooltip } from 'src/components/tooltip';
 import { WordPressLogoCircle } from 'src/components/wordpress-logo-circle';
 import { useSyncSites } from 'src/hooks/sync-sites';
-import { SYNC_OPTIONS, SyncOption, isSyncOption } from 'src/hooks/sync-sites/sync-option';
 import { useConfirmationDialog } from 'src/hooks/use-confirmation-dialog';
 import { useFeatureFlags } from 'src/hooks/use-feature-flags';
 import { useImportExport } from 'src/hooks/use-import-export';
@@ -25,9 +24,8 @@ import { cx } from 'src/lib/cx';
 import { getIpcApi } from 'src/lib/get-ipc-api';
 import { getLocalizedLink } from 'src/lib/get-localized-link';
 import { OpenSitesSyncSelector } from 'src/modules/content-tab-sync';
-import { SyncDialog } from 'src/modules/content-tab-sync/sync-dialog';
+import { SyncDialog } from 'src/modules/content-tab-sync/components/sync-dialog';
 import { useI18nLocale } from 'src/stores';
-import type { TreeNode } from 'src/components/tree-view';
 import type { SyncSite } from 'src/hooks/use-fetch-wpcom-sites/types';
 
 interface ConnectedSiteSection {
@@ -36,66 +34,6 @@ interface ConnectedSiteSection {
 	provider: 'wpcom';
 	connectedSites: SyncSite[];
 }
-
-export type SyncOptionsWithSelections = {
-	optionsToSync: SyncOption[];
-	specificSelections?: {
-		plugins?: string[];
-		themes?: string[];
-		uploads?: string[];
-	};
-};
-
-export const convertTreeToOptionsToSync = ( tree: TreeNode[] ): SyncOptionsWithSelections => {
-	const optionsToSync: SyncOption[] = [];
-	let specificSelections: SyncOptionsWithSelections[ 'specificSelections' ] = undefined;
-
-	const isAll = tree.every( ( node ) => node.checked );
-	if ( isAll ) {
-		optionsToSync.push( SYNC_OPTIONS.all );
-	} else {
-		const isDatabaseSelected = tree.find( ( node ) => node.id === SYNC_OPTIONS.sqls )?.checked;
-
-		if ( isDatabaseSelected ) {
-			optionsToSync.push( SYNC_OPTIONS.sqls );
-		}
-
-		const filesAndFolders = tree.find( ( node ) => node.id === 'filesAndFolders' )?.children || [];
-		const wpContent = filesAndFolders.find( ( node ) => node.id === 'wp-content' )?.children || [];
-
-		wpContent.forEach( ( item ) => {
-			if ( ! isSyncOption( item.id ) ) {
-				return;
-			}
-
-			if ( item.checked || item.indeterminate ) {
-				optionsToSync.push( item.id );
-			}
-
-			if (
-				item.children &&
-				[ SYNC_OPTIONS.plugins, SYNC_OPTIONS.themes, SYNC_OPTIONS.uploads ].includes(
-					item.id as 'plugins' | 'themes' | 'uploads'
-				)
-			) {
-				const selectedItems = item.children
-					.filter( ( child ) => child.checked )
-					.map( ( child ) => child.id );
-				if ( selectedItems.length > 0 && selectedItems.length < item.children.length ) {
-					specificSelections = {
-						...specificSelections,
-						[ item.id ]: selectedItems,
-					};
-				}
-			}
-		} );
-	}
-
-	return {
-		optionsToSync,
-		specificSelections,
-	};
-};
 
 const SyncConnectedSiteControls = ( {
 	connectedSite,
