+++ src/modules/content-tab-sync/components/sync-dialog/hooks/use-default-sync-tree.tsx
@@ -1,7 +1,7 @@
 import { useI18n } from '@wordpress/react-i18n';
 import { useMemo } from 'react';
 import { TreeNode } from 'src/components/tree-view';
-import { SYNC_OPTIONS } from 'src/hooks/sync-sites/sync-option';
+import { SYNC_OPTIONS } from 'src/constants';
 
 export const useDefaultSyncTree = (): TreeNode[] => {
 	const { __ } = useI18n();
