+++ src/modules/content-tab-sync/images/sync-tab-image.tsx
