--- src/modules/content-tab-sync/sync-dialog/types.ts
@@ -1,3 +1,0 @@
-import { GRANULAR_SYNC_FOLDERS } from 'src/modules/content-tab-sync/sync-dialog/constants';
-
-export type GranularSyncFolders = ( typeof GRANULAR_SYNC_FOLDERS )[ number ];
