+++ src/types.ts
@@ -1,0 +1,3 @@
+import { SYNC_OPTIONS } from 'src/constants';
+
+export type SyncOption = keyof typeof SYNC_OPTIONS;
